# backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.BookResponse:
    properties:
      authors:
        items:
          type: string
        type: array
      chapters:
        items:
          $ref: '#/definitions/handlers.ChapterSummary'
        type: array
      description:
        type: string
      id:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  handlers.ChapterSummary:
    properties:
      id:
        type: integer
      order:
        type: integer
      title:
        type: string
    type: object
  handlers.CreateBookRequest:
    properties:
      authors:
        items:
          type: integer
        minItems: 1
        type: array
      description:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - authors
    - description
    - title
    type: object
  handlers.CreateBookResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  handlers.CreateChapterRequestBody:
    properties:
      content:
        type: string
      order:
        type: integer
      title:
        type: string
    required:
    - content
    - order
    - title
    type: object
  handlers.CreateChapterResponse:
    properties:
      book_id:
        type: integer
      content:
        type: string
      id:
        type: integer
      order:
        type: integer
      title:
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      message:
        type: string
      success:
        example: false
        type: boolean
    type: object
  handlers.GetChapterResponse:
    properties:
      book_id:
        type: integer
      content:
        type: string
      id:
        type: integer
      order:
        type: integer
      title:
        type: string
    type: object
  handlers.RegisterUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  handlers.RegisterUserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  handlers.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
  handlers.TagResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
info:
  contact: {}
  description: Doko backend API documentation.
  title: Doko API
  version: "0.1"
paths:
  /api/book/:
    post:
      consumes:
      - application/json
      description: Create a new book with authors and optional tags.
      parameters:
      - description: Book payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.CreateBookResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create book
      tags:
      - book
  /api/book/{id}:
    delete:
      description: Delete a book by ID.
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete book
      tags:
      - book
    get:
      description: Get a book by ID.
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.BookResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get book
      tags:
      - book
  /api/book/{id}/chapter:
    post:
      consumes:
      - application/json
      description: Create a chapter under a book.
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Chapter payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateChapterRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.CreateChapterResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create chapter
      tags:
      - book
  /api/book/chapter/{id}:
    get:
      description: Get a chapter by ID.
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.GetChapterResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get chapter
      tags:
      - book
  /api/book/recent/{limit}:
    get:
      description: Get recently created books.
      parameters:
      - description: Limit
        in: path
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.BookResponse'
                  type: array
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get recent books
      tags:
      - book
  /api/tag/:
    get:
      description: Get all tags.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.TagResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get tags
      tags:
      - tag
  /api/user/{id}:
    delete:
      description: Delete a user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete user
      tags:
      - user
  /api/user/register:
    post:
      consumes:
      - application/json
      description: Create a new user account.
      parameters:
      - description: Register payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.RegisterUserResponse'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Register user
      tags:
      - user
swagger: "2.0"